fix(image-loader): use tempDirectory on iOS (#39)
* fix(image-loader): use tempDirectory on iOS

The tempDirectory can be accessed from the WKWebView as well; no need
for base64!

* Revert "fix(image-loader): use tempDirectory on iOS"

This reverts commit bea58390f416fbf5f42fdcd0a895573524eba87d.

* Specify the imageReturnType (base64/uri)

If 'uri' is used the files will be copied to the temporary directory if
the WKWebView engine is used. Files in this directory are not
persistent, but accessible from within the WebView.
If the files do not exist any longer they will automatically be copied
again!